requirements.txt
streamlit==0.87.2
pandas==1.3.2
plotly==5.2.2
matplotlib==3.4.3
Pillow==8.3.1





